refactor(data-model,runner)!: move the `data:` URI codec into `data-model`
`data-uri-codec.ts` needs nothing from the runner -- that was the whole
point of the module split -- so it moves to the package it actually
belongs to, exported as `@commonfabric/data-model/data-uri-codec`. Its
pure-codec unit tests (`valueFromDataUriPayloadText`, `valueFromDataUri`)
move with it, plus direct tests for the mint half and the predicates;
the runner keeps the cell-side `dataUriFromValueWithResolvedLinks` suite.

The one signature change: the codec no longer imports `URI` from
`memory/interface` (data-model must not depend on memory). A local
`UriString` -- structurally identical to `URI`, so interchangeable at
every call site -- takes its place, used only in that file.
